feat : 간단한 리팩토링
