+++ src/components/ui/TradePassword.tsx
@@ -1,0 +1,3 @@
+export default function TradePassword() {
+  return <div>dfkslfjs</div>
+}
